# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebBanSachOnline API",
    "version": "1.0.0",
    "description": "Tài liệu API cho dự án WebBanSachOnline"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/me": {
      "get": {
        "summary": "Lấy thông tin tài khoản đang đăng nhập",
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trả về { user, profile } của người dùng hiện tại"
          },
          "401": {
            "description": "Chưa đăng nhập"
          }
        }
      },
      "patch": {
        "summary": "Cập nhật thông tin cá nhân",
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "Họ và tên (2-100 ký tự, Unicode)"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Địa chỉ email mới"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "Số điện thoại (định dạng VN)"
                  },
                  "default_address": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Địa chỉ giao hàng mặc định"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công, trả về { profile }"
          },
          "400": {
            "description": "Lỗi validate dữ liệu hoặc lỗi trùng email"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc token không hợp lệ"
          },
          "403": {
            "description": "Lỗi quyền hạn (RLS)"
          },
          "404": {
            "description": "Không tìm thấy hồ sơ người dùng"
          }
        }
      }
    },
    "/me/avatar": {
      "post": {
        "summary": "Cập nhật ảnh đại diện",
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "Dữ liệu hình ảnh dạng Base64 (data:image/...;base64,...)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "400": {
            "description": "Lỗi dữ liệu hoặc tải lên thất bại"
          }
        }
      },
      "delete": {
        "summary": "Xóa ảnh đại diện",
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
    "/me/change-password": {
      "post": {
        "summary": "Thay đổi mật khẩu",
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "401": {
            "description": "Mật khẩu hiện tại sai"
          }
        }
      }
    },
...
  }
}